refactor(server-cloudflare): rename src files for clarity
- factory.ts → env.ts (file defines env types, not a factory pattern)
- auth/better-auth.ts → auth/server.ts (name the concept, not the library)
- ai/chat.ts → proxy/chat.ts (consolidate both proxy handlers)
- proxy/handler.ts → proxy/passthrough.ts (distinguish from chat proxy)
