+++ packages/server-cloudflare/src/auth.ts
@@ -2,7 +2,7 @@
 // Reads DATABASE_URL from .env / environment — no Cloudflare bindings needed.
 // Spreads sharedAuthConfig so the CLI schema always matches the runtime worker.
 import { betterAuth } from 'better-auth';
-import { sharedAuthConfig } from './auth/better-auth';
+import { sharedAuthConfig } from './auth/server';
 
 export const auth = betterAuth({
 	...sharedAuthConfig,
